+++ src/components/Checklist.jsx
@@ -9,7 +9,7 @@ import {
   fetchStockHistory,
   addStock,
 } from "../../utils/supabaseFunctions";
-import { saveAs } from "file-saver";
+import { exportToCSV } from "../../utils/exportToCSV";
 import Header from "./Header";
 
 const Checklist = () => {
@@ -187,23 +187,6 @@ const Checklist = () => {
   // モーダルで「いいえ」がクリックされたとき
   const handleModalNo = () => {
     setShowModal(false);
-  };
-
-  const exportToCSV = (stocks, items) => {
-    let csvContent = "項目名,在庫数,単位\n"; // ヘッダーを追加
-
-    for (const itemId in stocks) {
-      const item = items.find((item) => item.id.toString() === itemId);
-      if (!item) continue;
-
-      for (const unit in stocks[itemId]) {
-        const unitName = unit === "unit1" ? item.unit1 : item.unit2;
-        csvContent += `${item.item},${stocks[item.id][unit]},${unitName}\n`;
-      }
-    }
-
-    const blob = new Blob([csvContent], { type: "text/csv;charset=utf-8;" });
-    saveAs(blob, "stocks.csv");
   };
 
   // アコーディオンの開閉状態を切り替え
